+++ pybamm/simulation.py
@@ -91,11 +91,16 @@ class Simulation:
                             * self._parameter_values["Nominal cell capacity [A.h]"]
                         }
                     )
-
-            self._unprocessed_model = model
-            self.model = model
         else:
-            self.set_up_experiment(model, experiment)
+            if not isinstance(experiment, pybamm.Experiment):
+                raise TypeError("experiment must be a pybamm `Experiment` instance")
+
+            self.operating_mode = "with experiment"
+            # Save the experiment
+            self.experiment = experiment.copy()
+
+        self._unprocessed_model = model
+        self.model = model
 
         self.geometry = geometry or self.model.default_geometry
         self.submesh_types = submesh_types or self.model.default_submesh_types
@@ -120,7 +125,7 @@ class Simulation:
 
             warnings.filterwarnings("ignore")
 
-    def set_up_experiment(self, model, experiment):
+    def set_up_and_parameterise_experiment(self):
         """
         Set up a simulation to run with an experiment. This creates a dictionary of
         inputs (current/voltage/power, running time, stopping condition) for each
@@ -130,122 +135,70 @@ class Simulation:
         This needs to be done here and not in the Experiment class because the nominal
         cell capacity (from the parameters) is used to convert C-rate to current.
         """
-        self.operating_mode = "with experiment"
-
-        if not isinstance(experiment, pybamm.Experiment):
-            raise TypeError("experiment must be a pybamm `Experiment` instance")
-
-        # Save the experiment
-        self.experiment = experiment
-        # Create a new submodel for each set of operating conditions and update
-        # parameters and events accordingly
-        self._experiment_inputs = []
-        self._experiment_times = []
-        for op, events in zip(experiment.operating_conditions, experiment.events):
-            operating_inputs = {}
-            op_units = op["electric"][1]
-            if op["dc_data"] is not None:
+        experiment = self.experiment
+        model = self.model
+        # Update experiment using parameters such as timescale and capacity
+        timescale = self._parameter_values.evaluate(model.timescale)
+        capacity = self._parameter_values["Nominal cell capacity [A.h]"]
+        for op_conds in experiment.operating_conditions:
+            op_type = op_conds["type"]
+            if op_conds["dc_data"] is not None:
                 # If operating condition includes a drive cycle, define the interpolant
-                timescale = self._parameter_values.evaluate(model.timescale)
                 drive_cycle_interpolant = pybamm.Interpolant(
-                    op["dc_data"][:, 0],
-                    op["dc_data"][:, 1],
+                    op_conds["dc_data"][:, 0],
+                    op_conds["dc_data"][:, 1],
                     timescale * (pybamm.t - pybamm.InputParameter("start time")),
                 )
-                if op_units == "A":
-                    operating_inputs.update(
-                        {
-                            "Current switch": 1,
-                            "Current input [A]": drive_cycle_interpolant,
-                        }
-                    )
-                if op_units == "V":
-                    operating_inputs.update(
-                        {
-                            "Voltage switch": 1,
-                            "Voltage input [V]": drive_cycle_interpolant,
-                        }
-                    )
-                if op_units == "W":
-                    operating_inputs.update(
-                        {"Power switch": 1, "Power input [W]": drive_cycle_interpolant}
-                    )
+                if op_type == "current":
+                    op_conds["Current input [A]"] = drive_cycle_interpolant
+                if op_type == "voltage":
+                    op_conds["Voltage input [V]"] = drive_cycle_interpolant
+                if op_type == "power":
+                    op_conds["Power input [W]"] = drive_cycle_interpolant
             else:
-                if op_units in ["A", "C"]:
-                    capacity = self._parameter_values["Nominal cell capacity [A.h]"]
-                    if op_units == "A":
-                        I = op["electric"][0]
-                        Crate = I / capacity
-                    else:
-                        # Scale C-rate with capacity to obtain current
-                        Crate = op["electric"][0]
-                        I = Crate * capacity
-                    if len(op["electric"]) == 4:
-                        # Update inputs for CCCV
-                        op_units = "CCCV"  # change to CCCV
-                        V = op["electric"][2]
-                        operating_inputs.update(
-                            {
-                                "CCCV switch": 1,
-                                "Current input [A]": I,
-                                "Voltage input [V]": V,
-                            }
-                        )
-                    else:
-                        # Update inputs for constant current
-                        operating_inputs.update(
-                            {"Current switch": 1, "Current input [A]": I}
-                        )
-                elif op_units == "V":
-                    # Update inputs for constant voltage
-                    V = op["electric"][0]
-                    operating_inputs.update(
-                        {"Voltage switch": 1, "Voltage input [V]": V}
-                    )
-                elif op_units == "W":
-                    # Update inputs for constant power
-                    P = op["electric"][0]
-                    operating_inputs.update({"Power switch": 1, "Power input [W]": P})
-
-            # Update period
-            operating_inputs["period"] = op["period"]
+                if op_type == "C-rate":
+                    Crate = op_conds.pop("C-rate input [-]")
+                    op_conds["type"] = "current"
+                    op_conds["Current input [A]"] = Crate * capacity
+                elif op_type == "current":
+                    Crate = op_conds["Current input [A]"] / capacity
 
             # Update events
-            if events is None:
-                pass
-            elif events[1] in ["A", "C"]:
-                # update current cut-off, make voltage a value that won't be hit
-                if events[1] == "A":
-                    I = events[0]
-                else:
+            events = op_conds.pop("events")
+            if events is not None:
+                event_type = events.pop("type")
+                if event_type == "C-rate":
                     # Scale C-rate with capacity to obtain current
-                    capacity = self._parameter_values["Nominal cell capacity [A.h]"]
-                    I = events[0] * capacity
-                operating_inputs.update({"Current cut-off [A]": I})
-            elif events[1] == "V":
-                # update voltage cut-off, make current a value that won't be hit
-                V = events[0]
-                operating_inputs.update({"Voltage cut-off [V]": V})
+                    events["Current input [A]"] = (
+                        events.pop("C-rate input [-]") * capacity
+                    )
+                # Update the dictionary of operating conditions, replacing
+                # "xxx input [unit]" with "xxx cut-off [unit]"
+                op_conds.update(
+                    {
+                        key.replace("input", "cut-off"): value
+                        for key, value in events.items()
+                    }
+                )
 
-            self._experiment_inputs.append(operating_inputs)
             # Add time to the experiment times
-            dt = op["time"]
+            dt = op_conds["time"]
             if dt is None:
-                if op_units in ["A", "C", "CCCV"]:
+                if op_conds["type"] in ["current", "CCCV"]:
                     # Current control: max simulation time: 3 * max simulation time
                     # based on C-rate
                     dt = 3 / abs(Crate) * 3600  # seconds
-                    if op_units == "CCCV":
+                    if op_conds["type"] == "CCCV":
                         dt *= 5  # 5x longer for CCCV
                 else:
                     # max simulation time: 1 day
                     dt = 24 * 3600  # seconds
-            self._experiment_times.append(dt)
+            op_conds["time"] = dt
 
         # Set up model for experiment
-        self.set_up_model_for_experiment(model)
+        self.set_up_and_parameterise_model_for_experiment()
 
-    def set_up_model_for_experiment(self, model):
+    def set_up_and_parameterise_model_for_experiment(self):
         """
         Set up self.model to be able to run the experiment (new version).
         In this version, a new model is created for each step.
@@ -253,17 +206,14 @@ class Simulation:
         This increases set-up time since several models to be processed, but
         reduces simulation time since the model formulation is efficient.
         """
-        self.op_conds_to_model_and_param = {}
-        for op_cond, op_inputs in zip(
-            self.experiment.operating_conditions, self._experiment_inputs
-        ):
-            # Create model for this operating condition if it has not already been seen
-            # before
-            if op_cond["electric"] not in self.op_conds_to_model_and_param:
-                if op_inputs.get("Current switch") == 1:
-                    # Current control
-                    # Make a new copy of the model (we will update events later))
-                    new_model = model.new_copy()
+        self.op_type_to_model = {}
+        self.op_string_to_model = {}
+        for op in self.experiment.operating_conditions:
+            # Create model for this operating condition type (current/voltage/power)
+            # if it has not already been seen before
+            if op["type"] not in self.op_type_to_model:
+                if op["type"] == "current":
+                    new_model, submodel = self.model, None
                 else:
                     # Voltage or power control
                     # Create a new model where the current density is now a variable
@@ -272,176 +222,142 @@ class Simulation:
                     # FunctionControl submodel
                     # check which kind of external circuit model we need (differential
                     # or algebraic)
-                    if op_inputs.get("CCCV switch") == 1:
-                        control = "differential with max"
-                    else:
-                        control = "algebraic"
-                    # create the FunctionControl submodel and extract variables
-                    external_circuit_variables = (
-                        pybamm.external_circuit.FunctionControl(
-                            model.param, None, model.options, control=control
-                        ).get_fundamental_variables()
-                    )
+                    if op["type"] == "voltage":
+                        submodel_class = pybamm.external_circuit.VoltageFunctionControl
+                    elif op["type"] == "power":
+                        submodel_class = pybamm.external_circuit.PowerFunctionControl
+                    elif op["type"] == "CCCV":
+                        submodel_class = pybamm.external_circuit.CCCVFunctionControl
 
-                    # Perform the replacement
-                    symbol_replacement_map = {
-                        model.variables[name]: variable
-                        for name, variable in external_circuit_variables.items()
-                    }
-                    # Don't replace initial conditions, as these should not contain
-                    # Variable objects
-                    replacer = pybamm.SymbolReplacer(
-                        symbol_replacement_map, process_initial_conditions=False
-                    )
-                    new_model = replacer.process_model(model, inplace=False)
+                    new_model = self.model.new_copy()
+                    # Build the new submodel and update the model with it
+                    submodel = submodel_class(new_model.param, new_model.options)
+                    variables = new_model.variables
+                    submodel.variables = submodel.get_fundamental_variables()
+                    variables.update(submodel.variables)
+                    submodel.variables.update(submodel.get_coupled_variables(variables))
+                    variables.update(submodel.variables)
+                    submodel.set_rhs(variables)
+                    submodel.set_algebraic(variables)
+                    submodel.set_initial_conditions(variables)
+                    new_model.rhs.update(submodel.rhs)
+                    new_model.algebraic.update(submodel.algebraic)
+                    new_model.initial_conditions.update(submodel.initial_conditions)
 
-                    # Update the rhs or algebraic equation and initial conditions for
-                    # FunctionControl
-                    # This creates a differential or algebraic equation for the current
-                    # to allow current, voltage, or power control, together with the
-                    # appropriate guess for the initial condition.
-                    # External circuit submodels are always equations on the current
-                    # The external circuit function should fix either the current, or
-                    # the voltage, or a combination (e.g. I*V for power control)
-                    i_cell = new_model.variables["Current density variable"]
-                    new_model.initial_conditions[
-                        i_cell
-                    ] = new_model.param.current_with_time
+                self.op_type_to_model[op["type"]] = (new_model, submodel)
 
-                    # add current events to the model
-                    if "Current cut-off [A]" in op_inputs:
-                        if op_inputs.get("CCCV switch") == 1:
-                            # for the CCCV model we need to make sure that the current
-                            # cut-off is only reached at the end of the CV phase
-                            # Current is negative for a charge so this event will be
-                            # negative until it is zero
-                            # So we take away a large number times a heaviside switch
-                            # for the CV phase to make sure that the event can only be
-                            # hit during CV
-                            new_model.events.append(
-                                pybamm.Event(
-                                    "Current cut-off (CCCV) [A] [experiment]",
-                                    -new_model.variables["Current [A]"]
-                                    - abs(pybamm.InputParameter("Current cut-off [A]"))
-                                    + 1e4
-                                    * (
-                                        new_model.variables["Battery voltage [V]"]
-                                        < (
-                                            pybamm.InputParameter("Voltage input [V]")
-                                            - 1e-4
-                                        )
-                                    ),
-                                )
-                            )
-                        else:
-                            # current event
-                            new_model.events.extend(
-                                [
-                                    pybamm.Event(
-                                        "Current cut-off [A] [experiment]",
-                                        abs(new_model.variables["Current [A]"])
-                                        - pybamm.InputParameter("Current cut-off [A]"),
-                                    ),
-                                ]
-                            )
-                    if op_inputs.get("Voltage switch") == 1:
-                        new_model.algebraic[
-                            i_cell
-                        ] = pybamm.external_circuit.VoltageFunctionControl(
-                            new_model.param, model.options
-                        ).constant_voltage(
-                            new_model.variables
-                        )
-                    elif op_inputs.get("Power switch") == 1:
-                        new_model.algebraic[
-                            i_cell
-                        ] = pybamm.external_circuit.PowerFunctionControl(
-                            new_model.param, new_model.options, control="algebraic"
-                        ).constant_power(
-                            new_model.variables
-                        )
-                    elif op_inputs.get("CCCV switch") == 1:
-                        new_model.rhs[
-                            i_cell
-                        ] = pybamm.external_circuit.CCCVFunctionControl(
-                            new_model.param, new_model.options
-                        ).cccv(
-                            new_model.variables
-                        )
-
-                # add voltage events to the model
-                if "Voltage cut-off [V]" in op_inputs:
-                    # The voltage event should be positive at the start of charge/
-                    # discharge. We use the sign of the current or power input to
-                    # figure out whether the voltage event is greater than the starting
-                    # voltage (charge) or less (discharge) and set the sign of the
-                    # event accordingly
-                    if op_inputs.get("Power switch") == 1:
-                        inp = op_inputs["Power input [W]"]
-                    else:
-                        inp = op_inputs["Current input [A]"]
-                    sign = np.sign(inp)
-                    if sign > 0:
-                        name = "Discharge"
-                    else:
-                        name = "Charge"
-                    if sign != 0:
-                        # Event should be positive at initial conditions for both
-                        # charge and discharge
-                        new_model.events.append(
-                            pybamm.Event(
-                                f"{name} voltage cut-off [V] [experiment]",
-                                sign
-                                * (
-                                    new_model.variables["Battery voltage [V]"]
-                                    - pybamm.InputParameter("Voltage cut-off [V]")
-                                ),
-                            )
-                        )
-
-                # Keep the min and max voltages as safeguards but add some tolerances
-                # so that they are not triggered before the voltage limits in the
-                # experiment
-                for i, event in enumerate(new_model.events):
-                    if event.name in ["Minimum voltage", "Maximum voltage"]:
-                        new_model.events[i] = pybamm.Event(
-                            event.name, event.expression + 1, event.event_type
-                        )
-
+            if op["string"] not in self.op_string_to_model:
+                model, submodel = self.op_type_to_model[op["type"]]
+                # Create a new model for this operating condition, since we will update
+                # the events differently (based on parameter values and inputs) for
+                # different models of the same type (current/voltage/power)
+                new_model = model.new_copy()
+                self.update_new_model_events(new_model, op)
                 # Update parameter values
                 new_parameter_values = self.parameter_values.copy()
-                if op_inputs.get("Current switch") == 1:
-                    new_parameter_values.update(
-                        {"Current function [A]": op_inputs["Current input [A]"]}
-                    )
-                elif op_inputs.get("Voltage switch") == 1:
-                    new_parameter_values.update(
-                        {
-                            "Voltage function [V]": op_inputs["Voltage input [V]"]
-                            / model.param.n_cells
-                        },
-                        check_already_exists=False,
-                    )
-                elif op_inputs.get("Power switch") == 1:
-                    new_parameter_values.update(
-                        {"Power function [W]": op_inputs["Power input [W]"]},
-                        check_already_exists=False,
-                    )
-                elif op_inputs.get("CCCV switch") == 1:
-                    new_parameter_values.update(
-                        {
-                            "Current function [A]": op_inputs["Current input [A]"],
-                            "Voltage function [V]": op_inputs["Voltage input [V]"]
-                            / model.param.n_cells,
-                        },
-                        check_already_exists=False,
-                    )
-
-                self.op_conds_to_model_and_param[op_cond["electric"]] = (
-                    new_model,
-                    new_parameter_values,
+                experiment_parameter_values = self.get_experiment_parameter_values(op)
+                new_parameter_values.update(
+                    experiment_parameter_values, check_already_exists=False
                 )
-        self.model = model
+                # Set the "current function" to be the variable defined in the
+                if submodel is not None:
+                    new_parameter_values["Current function [A]"] = submodel.variables[
+                        "Current density variable"
+                    ] * abs(model.param.I_typ)
+                parameterised_model = new_parameter_values.process_model(
+                    new_model, inplace=False
+                )
+                self.op_string_to_model[op["string"]] = parameterised_model
+
+    def update_new_model_events(self, new_model, op):
+        if "Current cut-off [A]" in op:
+            if op["type"] == "CCCV":
+                # for the CCCV model we need to make sure that the current
+                # cut-off is only reached at the end of the CV phase
+                # Current is negative for a charge so this event will be
+                # negative until it is zero
+                # So we take away a large number times a heaviside switch
+                # for the CV phase to make sure that the event can only be
+                # hit during CV
+                new_model.events.append(
+                    pybamm.Event(
+                        "Current cut-off (CCCV) [A] [experiment]",
+                        -new_model.variables["Current [A]"]
+                        - abs(pybamm.InputParameter("Current cut-off [A]"))
+                        + 1e4
+                        * (
+                            new_model.variables["Battery voltage [V]"]
+                            < (pybamm.InputParameter("Voltage input [V]") - 1e-4)
+                        ),
+                    )
+                )
+            else:
+                new_model.events.append(
+                    pybamm.Event(
+                        "Current cut-off [A] [experiment]",
+                        abs(new_model.variables["Current [A]"])
+                        - pybamm.InputParameter("Current cut-off [A]"),
+                    )
+                )
+
+        # add voltage events to the model
+        if "Voltage cut-off [V]" in op:
+            # The voltage event should be positive at the start of charge/
+            # discharge. We use the sign of the current or power input to
+            # figure out whether the voltage event is greater than the starting
+            # voltage (charge) or less (discharge) and set the sign of the
+            # event accordingly
+            if op["type"] == "power":
+                inp = op["Power input [W]"]
+            else:
+                inp = op["Current input [A]"]
+            sign = np.sign(inp)
+            if sign > 0:
+                name = "Discharge"
+            else:
+                name = "Charge"
+            if sign != 0:
+                # Event should be positive at initial conditions for both
+                # charge and discharge
+                new_model.events.append(
+                    pybamm.Event(
+                        f"{name} voltage cut-off [V] [experiment]",
+                        sign
+                        * (
+                            new_model.variables["Battery voltage [V]"]
+                            - pybamm.InputParameter("Voltage cut-off [V]")
+                        ),
+                    )
+                )
+
+        # Keep the min and max voltages as safeguards but add some tolerances
+        # so that they are not triggered before the voltage limits in the
+        # experiment
+        for i, event in enumerate(new_model.events):
+            if event.name in ["Minimum voltage", "Maximum voltage"]:
+                new_model.events[i] = pybamm.Event(
+                    event.name, event.expression + 1, event.event_type
+                )
+
+    def get_experiment_parameter_values(self, op):
+        experiment_parameter_values = {}
+        if op["type"] == "current":
+            experiment_parameter_values.update(
+                {"Current function [A]": op["Current input [A]"]}
+            )
+        if op["type"] == "CCCV":
+            experiment_parameter_values.update(
+                {"CCCV current function [A]": op["Current input [A]"]}
+            )
+        if op["type"] in ["voltage", "CCCV"]:
+            experiment_parameter_values.update(
+                {"Voltage function [V]": op["Voltage input [V]"]}
+            )
+        if op["type"] == "power":
+            experiment_parameter_values.update(
+                {"Power function [W]": op["Power input [W]"]}
+            )
+        return experiment_parameter_values
 
     def set_parameters(self):
         """
@@ -486,17 +402,6 @@ class Simulation:
                 "Initial concentration in positive electrode [mol.m-3]": y * c_p_max,
             }
         )
-        # For experiments also update the following
-        if hasattr(self, "op_conds_to_model_and_param"):
-            for key, (model, param) in self.op_conds_to_model_and_param.items():
-                param.update(
-                    {
-                        "Initial concentration in negative electrode [mol.m-3]": x
-                        * c_n_max,
-                        "Initial concentration in positive electrode [mol.m-3]": y
-                        * c_p_max,
-                    }
-                )
         # Save solved initial SOC in case we need to re-build the model
         self._built_initial_soc = initial_soc
         return c_n_init, c_p_init
@@ -535,14 +440,19 @@ class Simulation:
                 self._model_with_set_params, inplace=False, check_model=check_model
             )
 
-    def build_for_experiment(self, check_model=True):
+    def build_for_experiment(self, check_model=True, initial_soc=None):
         """
         Similar to :meth:`Simulation.build`, but for the case of simulating an
         experiment, where there may be several models to build
         """
+        if initial_soc is not None:
+            self.set_initial_soc(initial_soc)
+
         if self.op_conds_to_built_models:
             return
         else:
+            self.set_up_and_parameterise_experiment()
+
             # Can process geometry with default parameter values (only electrical
             # parameters change between parameter values)
             self._parameter_values.process_geometry(self._geometry)
@@ -551,21 +461,12 @@ class Simulation:
             self._disc = pybamm.Discretisation(self._mesh, self._spatial_methods)
             # Process all the different models
             self.op_conds_to_built_models = {}
-            processed_models = {}
-            for op_cond, (
-                unbuilt_model,
-                parameter_values,
-            ) in self.op_conds_to_model_and_param.items():
-                model_with_set_params = parameter_values.process_model(
-                    unbuilt_model, inplace=False
-                )
+            for op_cond, model_with_set_params in self.op_string_to_model.items():
                 # It's ok to modify the model with set parameters in place as it's
                 # not returned anywhere
                 built_model = self._disc.process_model(
                     model_with_set_params, inplace=True, check_model=check_model
                 )
-
-                processed_models[unbuilt_model] = built_model
                 self.op_conds_to_built_models[op_cond] = built_model
 
     def solve(
@@ -636,11 +537,8 @@ class Simulation:
         callbacks = pybamm.callbacks.setup_callbacks(callbacks)
         logs = {}
 
-        if initial_soc is not None:
-            c_n_init, c_p_init = self.set_initial_soc(initial_soc)
-
         if self.operating_mode in ["without experiment", "drive cycle"]:
-            self.build(check_model=check_model)
+            self.build(check_model=check_model, initial_soc=initial_soc)
             if save_at_cycles is not None:
                 raise ValueError(
                     "'save_at_cycles' option can only be used if simulating an "
@@ -719,7 +617,7 @@ class Simulation:
 
         elif self.operating_mode == "with experiment":
             callbacks.on_experiment_start(logs)
-            self.build_for_experiment(check_model=check_model)
+            self.build_for_experiment(check_model=check_model, initial_soc=initial_soc)
             if t_eval is not None:
                 pybamm.logger.warning(
                     "Ignoring t_eval as solution times are specified by the experiment"
@@ -805,13 +703,10 @@ class Simulation:
                 for step_num in range(1, cycle_length + 1):
                     # Use 1-indexing for printing cycle number as it is more
                     # human-intuitive
-                    exp_inputs = self._experiment_inputs[idx]
-                    dt = self._experiment_times[idx]
-                    op_conds_str = self.experiment.operating_conditions_strings[idx]
-                    op_conds_elec = self.experiment.operating_conditions[idx][
-                        "electric"
-                    ]
-                    model = self.op_conds_to_built_models[op_conds_elec]
+                    op_conds = self.experiment.operating_conditions[idx]
+                    dt = op_conds["time"]
+                    op_conds_str = op_conds["string"]
+                    model = self.op_conds_to_built_models[op_conds_str]
 
                     logs["step number"] = (step_num, cycle_length)
                     logs["step operating conditions"] = op_conds_str
@@ -820,11 +715,11 @@ class Simulation:
                     start_time = current_solution.t[-1]
                     kwargs["inputs"] = {
                         **user_inputs,
-                        **exp_inputs,
+                        **op_conds,
                         "start time": start_time,
                     }
                     # Make sure we take at least 2 timesteps
-                    npts = max(int(round(dt / exp_inputs["period"])) + 1, 2)
+                    npts = max(int(round(dt / op_conds["period"])) + 1, 2)
                     try:
                         step_solution = solver.step(
                             current_solution,
@@ -952,15 +847,6 @@ class Simulation:
 
             callbacks.on_experiment_end(logs)
 
-        # reset parameter values
-        if initial_soc is not None:
-            self.parameter_values.update(
-                {
-                    "Initial concentration in negative electrode [mol.m-3]": c_n_init,
-                    "Initial concentration in positive electrode [mol.m-3]": c_p_init,
-                }
-            )
-
         return self.solution
 
     def step(
@@ -1080,7 +966,6 @@ class Simulation:
     @model.setter
     def model(self, model):
         self._model = copy.copy(model)
-        self._model_class = model.__class__
 
     @property
     def model_with_set_params(self):